Base Game
This is what we will be starting with. In BlueJ, right click and run the "void run()". You will asked to enter the game type and difficulty. Both of these are not functional so enter any string for them. You will then be asked what to do, enter "scan". This is the only mechanic that is currently working. It will create a 8x8 grid of stars with a small chance of Planets, Black Holes, Klingon's, and Star Bases also showing up. We will talk about the future of the game on Friday. -Matt